refactor(nest): move pure logic to src/core/
