+++ nestjs/src/auth/redis-access-token.store.ts
@@ -1,7 +1,7 @@
 import { createHash } from 'node:crypto';
 import type Redis from 'ioredis';
-import { AccessTokenRecord } from './access-token-record';
-import { AccessTokenStore } from './access-token.store';
+import { AccessTokenRecord } from '@/core/auth/access-token-record';
+import { AccessTokenStore } from '@/core/auth/access-token.store';
 
 const sha256 = (s: string) => createHash('sha256').update(s).digest('hex');
