+++ nestjs/src/core/highlights/highlight-normalizer.ts
