+++ nestjs/src/core/highlights/snapshot-reader.ts
