+++ nestjs/src/core/members/member.entity.ts
