+++ nestjs/src/members/members.repository.ts
@@ -3,14 +3,15 @@ import { BetterSQLite3Database } from 'drizzle-orm/better-sqlite3';
 import type { Db } from '@/db/db.module';
 import * as schema from '@/db/schema';
 import { weavingUser } from '@/db/schema';
-import { Member } from './member.entity';
+import { Member } from '@/core/members/member.entity';
+import { MembersRepository as MembersRepositoryPort } from '@/core/members/members.repository';
 
 // The Db union (NodePgDatabase | BetterSQLite3Database) has incompatible .select() overloads
 // at the TypeScript level. We cast to the sqlite variant (which is the test-time driver) so
 // the compiler is happy; at runtime the pg driver exposes the same API surface.
 type AnyDb = BetterSQLite3Database<typeof schema>;
 
-export class MembersRepository {
+export class MembersRepository implements MembersRepositoryPort {
   constructor(private readonly db: Db) {}
 
   private get anyDb(): AnyDb {
