+++ nestjs/test/doubles/in-memory-snapshot-reader.ts
@@ -1,4 +1,4 @@
-import { SnapshotReader } from '@/highlights/snapshot-reader';
+import { SnapshotReader } from '@/core/highlights/snapshot-reader';
 
 export class InMemorySnapshotReader implements SnapshotReader {
   constructor(private snapshotsByDate: Record<string, unknown[] | Record<string, unknown>> = {}) {}
